+++ src/base/phased_discrete_rule_state_provider.cc
@@ -46,6 +46,34 @@ using api::rules::PhaseRing;
 using api::rules::PhaseRingBook;
 using api::rules::Rule;
 
+namespace {
+
+// Sets to `state_provider` the first value in each
+// DiscreteValueRule::states() in `discrete_value_rules` as the default
+// state.
+// @throws maliput::common::assertion_error When `state_provider` is
+//         nullptr.
+void PopulateDiscreteValueRuleStates(const std::map<Rule::Id, DiscreteValueRule>& discrete_value_rules,
+                                     maliput::PhasedDiscreteRuleStateProvider* state_provider) {
+  MALIPUT_THROW_UNLESS(state_provider != nullptr);
+  for (const auto& rule_id_rule : discrete_value_rules) {
+    state_provider->SetState(rule_id_rule.first, rule_id_rule.second.states().front(), std::nullopt, std::nullopt);
+  }
+}
+
+}  // namespace
+
+std::unique_ptr<PhasedDiscreteRuleStateProvider>
+PhasedDiscreteRuleStateProvider::GetDefaultPhasedDiscreteRuleStateProvider(
+    const maliput::api::rules::RoadRulebook* rulebook, const maliput::api::rules::PhaseRingBook* phase_ring_book,
+    const maliput::api::rules::PhaseProvider* phase_provider) {
+  auto state_provider =
+      std::make_unique<maliput::PhasedDiscreteRuleStateProvider>(rulebook, phase_ring_book, phase_provider);
+  const maliput::api::rules::RoadRulebook::QueryResults all_rules = rulebook->Rules();
+  PopulateDiscreteValueRuleStates(all_rules.discrete_value_rules, state_provider.get());
+  return state_provider;
+}
+
 PhasedDiscreteRuleStateProvider::PhasedDiscreteRuleStateProvider(const api::rules::RoadRulebook* rulebook,
                                                                  const PhaseRingBook* phase_ring_book,
                                                                  const PhaseProvider* phase_provider)