+++ src/plugin/maliput_plugin_manager.cc
@@ -84,12 +84,12 @@ void MaliputPluginManager::AddPlugin(const std::string& path_to_plugin) {
       (is_repeated ? "A new version of Plugin Id: {} was loaded." : "Plugin Id: {} was correctly loaded."), id);
 }
 
-std::vector<MaliputPlugin::Id> MaliputPluginManager::ListPlugins() {
-  std::vector<MaliputPlugin::Id> keys{};
+std::unordered_map<MaliputPlugin::Id, MaliputPluginType> MaliputPluginManager::ListPlugins() const {
+  std::unordered_map<MaliputPlugin::Id, MaliputPluginType> id_type{};
   for (const auto& plugin : plugins_) {
-    keys.push_back(plugin.first);
+    id_type.emplace(plugin.first, plugin.second->GetType());
   }
-  return keys;
+  return id_type;
 }
 
 }  // namespace plugin