+++ CHANGELOG.rst
@@ -2,6 +2,22 @@
 Changelog for package maliput
 ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
 
+1.0.9 (2022-11-28)
+------------------
+* Provides default populated discrete and range value rule state providers. (`#533 <https://github.com/maliput/maliput/issues/533>`_)
+* Adds a default populated ManualPhaseProvider. (`#530 <https://github.com/maliput/maliput/issues/530>`_)
+* Maliput Plugin: Adds interface for providing default parameters from the backends (`#532 <https://github.com/maliput/maliput/issues/532>`_)
+* Contributors: Franco Cipollone
+
+1.0.8 (2022-11-10)
+------------------
+* Brings range validator from maliput_malidrive. (`#529 <https://github.com/maliput/maliput/issues/529>`_)
+* Update triage.yml (`#526 <https://github.com/maliput/maliput/issues/526>`_)
+* Adds convenient test utility method. (`#525 <https://github.com/maliput/maliput/issues/525>`_)
+* Adds a test function for LaneEnds. (`#524 <https://github.com/maliput/maliput/issues/524>`_)
+* Adds IsLanePositionResultClose macro. (`#522 <https://github.com/maliput/maliput/issues/522>`_)
+* Contributors: Agustin Alba Chicar, Franco Cipollone
+
 1.0.7 (2022-09-14)
 ------------------
 * Modifies ToLanePosition and adds ToSegmentPosition. (`#521 <https://github.com/maliput/maliput/issues/521>`_)